#fix multiple sorting
