--- Infrastructures/Repositories/ISubUnitRepository.cs
@@ -1,7 +1,0 @@
-﻿using AdminHRM.Server.Entities;
-
-namespace AdminHRM.Server.Infrastructures;
-
-public interface ISubUnitRepository : IGenericRepository<SubUnit>
-{
-}
