+++ Infrastructures/Repositories/SubUnitRepository.cs
@@ -3,6 +3,10 @@ using AdminHRM.Server.Entities;
 
 namespace AdminHRM.Server.Infrastructures;
 
+public interface ISubUnitRepository : IGenericRepository<SubUnit>
+{
+}
+
 public class SubUnitRepository : GenericRepository<SubUnit>, ISubUnitRepository
 {
     public SubUnitRepository(HrmDbContext hrmDbContext) : base(hrmDbContext)
