+++ Services/Implements/EmployeeService.cs
@@ -12,15 +12,6 @@ public interface IEmployeeServive
     Task<bool?> EditEmployeeAsync(EmployeeDto employeeDto);
     Task<bool?> RemoveEmployeeDtosAsync(Guid id);
     Task<List<EmployeeDto>> SearchEmployeeDtosAsync(SearchEmployeeDto searchEmployeeDto);
-
-    /// <summary>
-    /// Get paging for PIM screen
-    /// </summary>
-    /// <param name="page"></param>
-    /// <param name="pageSize"></param>
-    /// <param name="sortFields"></param>
-    /// <param name="sortOrders"></param>
-    /// <returns></returns>
     Task<PagedResult<EmployeeDto>> GetPagedEmployeesAsync(int page, int pageSize, string[] sortFields, string[] sortOrders);
 
 }
@@ -119,11 +110,10 @@ public class EmployeeServive : IEmployeeServive
         {
             var totalEmployees = await _employeeRepository.CountAsync().ConfigureAwait(false);
             var employees = await _employeeRepository.GetPagedAsync(page, pageSize, sortFields, sortOrders).ConfigureAwait(false);
-            var employeeDtos = _mapper.Map<List<EmployeeDto>>(employees);
 
             return new PagedResult<EmployeeDto>
             {
-                Items = employeeDtos,
+                Items = employees,
                 TotalCount = totalEmployees,
                 PageIndex = page,
                 PageSize = pageSize,
